interactive_builder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
pydantic==2.5.0
pyyaml>=6.0
openai==1.51.0
anthropic==0.18.1
httpx==0.25.2